+++ release_notes.md
@@ -12,7 +12,9 @@
 
 ### ✨ New Functionality
 
-- 
+- (Beta) Add support for the SAP-internal Zero Trust Identity Service
+  - Add a new module `connectivity-ztis`.
+  - Add support for the credential-type `X509_ATTESTED` for all OAuth2 flows.
 
 ### 📈 Improvements